Algorithm.EqSat.Egraph

Description

Equality Graph data structure Heavily based on hegg (https:/github.comalt-romes/hegg by alt-romes)

Documentation

type EClassId = Int

type ClassIdMap = IntMap

type ENode = SRTree EClassId

type EData = ()

type EGraphST a = State EGraph a

type Cost = Int

type CostFun = SRTree Cost -> Cost

data EGraph

Constructors

EGraph

Fields

_canonicalMap :: ClassIdMap EClassId

_eNodeToEClass :: Map ENode EClassId

_eClass :: ClassIdMap EClass

_worklist :: [(EClassId, ENode)]

_analysis :: [(EClassId, ENode)]

_nextId :: Int

data EClass

Constructors

EClass

Fields

_eClassId :: Int

_eNodes :: Set ENode

_parents :: [(EClassId, ENode)]

_height :: Int

_info :: EClassData

data Consts

Constructors

NotConst

ParamIx Int

ConstVal Double

data Property

Constructors

Positive

Negative

NonZero

Real

data EClassData

Constructors

EData

Fields

_cost :: Cost

_best :: ENode

_consts :: Consts

worklist :: Lens' EGraph [(EClassId, ENode)]

nextId :: Lens' EGraph Int

eNodeToEClass :: Lens' EGraph (Map ENode EClassId)

eClass :: Lens' EGraph (ClassIdMap EClass)

canonicalMap :: Lens' EGraph (ClassIdMap EClassId)

analysis :: Lens' EGraph [(EClassId, ENode)]

parents :: Lens' EClass [(EClassId, ENode)]

info :: Lens' EClass EClassData

height :: Lens' EClass Int

eNodes :: Lens' EClass (Set ENode)

eClassId :: Lens' EClass Int

cost :: Lens' EClassData Cost

consts :: Lens' EClassData Consts

best :: Lens' EClassData ENode

add :: CostFun -> ENode -> EGraphST EClassId

adds a new or existing e-node (merging if necessary)

rebuild :: CostFun -> EGraphST ()

rebuilds the e-graph after inserting or merging e-classes

repair :: CostFun -> EClassId -> ENode -> EGraphST ()

repairs e-node by canonizing its children if the canonized e-node already exists in e-graph, merge the e-classes

repairAnalysis :: CostFun -> EClassId -> ENode -> EGraphST ()

repair the analysis of the e-class considering the new added e-node

merge :: CostFun -> EClassId -> EClassId -> EGraphST EClassId

merge to equivalnt e-classes

modifyEClass :: CostFun -> EClassId -> EGraphST EClassId

joinData :: EClassData -> EClassData -> EClassData

makeAnalysis :: CostFun -> ENode -> EGraphST EClassData

Calculate e-node data (constant values and cost)

fromTree :: CostFun -> Fix SRTree -> (EClassId, EGraph)

Creates an e-graph from an expression tree

fromTrees :: CostFun -> [Fix SRTree] -> ([EClassId], EGraph)

Builds an e-graph from multiple independent trees

fromTreeWith :: CostFun -> EGraph -> Fix SRTree -> (EClassId, EGraph)

insert new tree into an existing e-graph

findRootClasses :: EGraph -> [EClassId]

returns all the root e-classes (e-class without parents)

getExpressionFrom :: EGraph -> EClassId -> Fix SRTree

returns one expression rooted at e-class eId

getAllExpressionsFrom :: EGraph -> EClassId -> [Fix SRTree]

returns all expressions rooted at e-class eId

getRndExpressionFrom :: EGraph -> EClassId -> State StdGen (Fix SRTree)

returns a random expression rooted at e-class eId

calculateHeights :: EGraphST ()

update the heights of each e-class won't work if there's no root

canonical :: EClassId -> EGraphST EClassId

gets the canonical id of an e-class

canonical' :: EClassId -> EGraph -> EClassId

canonize :: ENode -> EGraphST ENode

canonize the e-node children

sequence' :: Applicative f => SRTree (f a) -> f (SRTree a)

children :: SRTree a -> [a]

replaceChildren :: [a] -> SRTree b -> SRTree a

getOperator :: SRTree a -> SRTree ()

getEClass :: EClassId -> EGraphST EClass

gets an e-class with id c

emptyGraph :: EGraph

returns an empty e-graph

calculateCost :: CostFun -> SRTree EClassId -> EGraphST Cost

calculates the cost of a node

calculateConsts :: SRTree EClassId -> EGraphST Consts

check whether an e-node evaluates to a const

combineConsts :: SRTree Consts -> Consts