Algorithm.EqSat.EqSatDB

Description

Pattern matching and rule application functions Heavily based on hegg (https:/github.comalt-romes/hegg by alt-romes)

Documentation

type DB = Map (SRTree ()) IntTrie

data IntTrie

Constructors

IntTrie

Fields

_keys :: Set EClassId

_trie :: IntMap IntTrie

data Pattern

Constructors

Fixed (SRTree Pattern)

VarPat Char

data Rule

Constructors

Pattern :=> Pattern infix 3

Pattern :==: Pattern infix 3

Rule :| Condition infixl 2

type Query = [Atom]

type Condition = Map ClassOrVar ClassOrVar -> EGraph -> Bool

type ClassOrVar = Either EClassId Int

data Atom

Constructors

Atom ClassOrVar (SRTree ClassOrVar)

unFixPat :: Pattern -> SRTree Pattern

createDB :: EGraph -> DB

createDB creates a database of patterns from an e-graph it simply calls addToDB for every pair (e-node, e-class id) from the e-graph.

addToDB :: ENode -> EClassId -> State DB ()

addToDB adds an e-node and e-class id to the database

trie :: EClassId -> IntMap IntTrie -> IntTrie

Creates a singleton trie from an e-class id

populate :: Maybe IntTrie -> [EClassId] -> Maybe IntTrie

Populates an IntTrie with a sequence of e-class ids

canonizeMap :: (Map ClassOrVar ClassOrVar, ClassOrVar) -> EGraphST (Map ClassOrVar ClassOrVar, ClassOrVar)

applyMatch :: CostFun -> Rule -> (Map ClassOrVar ClassOrVar, ClassOrVar) -> EGraphST ()

applyMergeOnlyMatch :: CostFun -> Rule -> (Map ClassOrVar ClassOrVar, ClassOrVar) -> EGraphST ()

target :: Rule -> Pattern

source :: Rule -> Pattern

getConditions :: Rule -> [Condition]

classOfENode :: CostFun -> Map ClassOrVar ClassOrVar -> Pattern -> EGraphST (Maybe EClassId)

gets the e-node of the target of the rule TODO: add consts and modify

isConst :: EClassId -> EGraphST Bool

reprPrat :: CostFun -> Map ClassOrVar ClassOrVar -> Pattern -> EGraphST EClassId

adds the target of the rule into the e-graph

isValidHeight :: (Map ClassOrVar ClassOrVar, ClassOrVar) -> EGraph -> Bool

isValidConditions :: Condition -> (Map ClassOrVar ClassOrVar, ClassOrVar) -> EGraph -> Bool

returns True if the condition of a rule is valid for that match

match :: DB -> Pattern -> [(Map ClassOrVar ClassOrVar, ClassOrVar)]

Returns the substitution rules for every match of the pattern source inside the e-graph.

compileToQuery :: Pattern -> (Query, ClassOrVar)

Returns a Query (list of atoms) of a pattern

getInt :: ClassOrVar -> Int

getElems :: SRTree a -> [a]

returns the list of the children values

genericJoin :: DB -> Query -> [Map ClassOrVar ClassOrVar]

Creates the substituion map for the pattern variables for each one of the matched subgraph

domainX :: DB -> ClassOrVar -> Query -> [ClassOrVar]

returns the e-class id for a certain variable that matches the pattern described by the atoms

intersectAtoms :: ClassOrVar -> DB -> Query -> [EClassId]

returns all e-class id that can matches this sequence of atoms

intersectTries :: ClassOrVar -> Map ClassOrVar EClassId -> IntTrie -> [ClassOrVar] -> Maybe (Set EClassId)

searches for the intersection of e-class ids that matches each part of the query. Returns Nothing if the intersection is empty.

var is the current variable being investigated xs is the map of ids being investigated and their corresponding e-class id trie is the current trie of the pattern (i:ids) sequence of root : children of the atom to investigate

updateVar :: ClassOrVar -> ClassOrVar -> Query -> Query

updates all occurrence of var with the new id x

isDiffFrom :: Int -> ClassOrVar -> Bool

checks whether two ClassOrVar are different only check if it is a pattern variable, else returns true

elemOfAtom :: ClassOrVar -> Atom -> Bool

checks if v is an element of an atom

orderedVars :: Query -> [ClassOrVar]

sorts the variables in a query by the most frequently occurring