Algorithm.SRTree.ConfidenceIntervals

Description

Functions to optimize the parameters of an expression.

Documentation

data PType

profile likelihood algorithms: Bates (classical), ODE (faster), Constrained (fastest) The Constrained approach returns only the endpoints.

Constructors

Bates

ODE

Constrained

data CIType

Confidence Interval using Laplace approximation or profile likelihood.

Constructors

Laplace BasicStats

Profile BasicStats [ProfileT]

data BasicStats

Basic stats of the data: covariance of parameters, correlation, standard errors

Constructors

MkStats

Fields

_cov :: SRMatrix

_corr :: SRMatrix

_stdErr :: PVector

data CI

a confience interval is composed of the point estimate (est_), lower bound (_lower_) and upper bound (upper_)

Constructors

CI

Fields

est_ :: Double

lower_ :: Double

upper_ :: Double

data ProfileT

Constructors

ProfileT

Fields

_taus :: PVector

_thetas :: SRMatrix

_opt :: Double

_tau2theta :: Double -> Double

_theta2tau :: Double -> Double

showCI :: Int -> CI -> String

printCI :: Int -> CI -> IO ()

paramCI :: CIType -> Int -> PVector -> Double -> [CI]

Calculates the confidence interval of the parameters using Laplace approximation or Profile likelihood

predictionCI :: CIType -> Distribution -> (SRMatrix -> PVector) -> (SRMatrix -> [PVector]) -> (CI -> PVector -> Fix SRTree -> (Double -> Double, Double)) -> SRMatrix -> Fix SRTree -> PVector -> Double -> [CI] -> [CI]

calculates the prediction confidence interval using Laplace approximation or profile likelihood.

inverseDist :: Floating p => Distribution -> p -> p

replaceParam0 :: Fix SRTree -> Fix SRTree -> Fix SRTree

evalVar :: PVector -> Fix SRTree -> Fix SRTree

calcTheta0 :: Distribution -> Fix SRTree -> Fix SRTree

getAllProfiles :: PType -> Distribution -> Maybe Double -> SRMatrix -> PVector -> Fix SRTree -> PVector -> PVector -> [CI] -> Double -> [ProfileT]

getProfile :: Distribution -> Maybe Double -> SRMatrix -> PVector -> Fix SRTree -> PVector -> Double -> Double -> Int -> Either PVector ProfileT

getProfileCnstr :: Distribution -> Maybe Double -> SRMatrix -> PVector -> Fix SRTree -> PVector -> Double -> Double -> Int -> Either PVector ProfileT

getEndPoint :: Distribution -> Maybe Double -> Array S Ix2 Double -> Array S Ix1 Double -> Fix SRTree -> Array S Ix1 Double -> Double -> Int -> Bool -> Double

getProfileODE :: Distribution -> Maybe Double -> SRMatrix -> PVector -> Fix SRTree -> PVector -> Double -> CI -> Double -> Int -> Either PVector ProfileT

rk :: (Double -> PVector -> PVector) -> (Double, PVector) -> Double -> (Double, PVector)

getStatsFromModel :: Distribution -> Maybe Double -> SRMatrix -> PVector -> Fix SRTree -> PVector -> BasicStats

createSplines :: PVector -> SRMatrix -> Double -> Double -> Int -> (Double -> Double, Double -> Double)

getCol :: Int -> SRMatrix -> PVector

sortOnFirst :: PVector -> PVector -> [(Double, Double)]

splinesSketches :: Double -> PVector -> PVector -> (Double -> Double) -> Double -> Double

approximateContour :: Int -> Int -> [ProfileT] -> Int -> Int -> Double -> [(Double, Double)]