Data.SRTree

Description

Expression tree for Symbolic Regression

Documentation

data SRTree val

Tree structure to be used with Symbolic Regression algorithms. This structure is a fixed point of a n-ary tree.

Constructors

Var Int
index of the variables

Param Int
index of the parameter

Const Double
constant value, can be converted to a parameter

Uni Function val
univariate function

Bin Op val val
binary operator

data Function

Supported functions

Constructors

Id

Abs

Sin

Cos

Tan

Sinh

Cosh

Tanh

ASin

ACos

ATan

ASinh

ACosh

ATanh

Sqrt

Cbrt

Square

Log

Exp

data Op

Supported operators

Constructors

Add

Sub

Mul

Div

Power

param :: Int -> Fix SRTree

create a tree with a single node representing a parameter

var :: Int -> Fix SRTree

create a tree with a single node representing a variable

arity :: Fix SRTree -> Int

Arity of the current node

getChildren :: Fix SRTree -> [Fix SRTree]

Get the children of a node. Returns an empty list in case of a leaf node.

countNodes :: Fix SRTree -> Int

Count the number of nodes in a tree.

countVarNodes :: Fix SRTree -> Int

Count the number of Var nodes

countConsts :: Fix SRTree -> Int

Count the number of const nodes

countParams :: Fix SRTree -> Int

Count the number of Param nodes

countOccurrences :: Int -> Fix SRTree -> Int

Count the occurrences of variable indexed as ix

deriveBy :: Bool -> Int -> Fix SRTree -> Fix SRTree

Creates the symbolic partial derivative of a tree by variable dx (if p is False) or parameter dx (if p is True).

deriveByVar :: Int -> Fix SRTree -> Fix SRTree

Symbolic derivative by a variable

deriveByParam :: Int -> Fix SRTree -> Fix SRTree

Symbolic derivative by a parameter

derivative :: Floating a => Function -> a -> a

forwardMode :: (Show a, Num a, Floating a) => Vector a -> Vector Double -> (Double -> a) -> Fix SRTree -> [a]

Calculates the numerical derivative of a tree using forward mode provided a vector of variable values xss, a vector of parameter values theta and a function that changes a Double value to the type of the variable values.

gradParams :: (Show a, Num a, Floating a) => Vector a -> Vector Double -> (Double -> a) -> Fix SRTree -> (a, [a])

The function gradParams calculates the numerical gradient of the tree and evaluates the tree at the same time. It assumes that each parameter has a unique occurrence in the expression. This should be significantly faster than forwardMode.

evalFun :: Floating a => Function -> a -> a

evalOp :: Floating a => Op -> a -> a -> a

inverseFunc :: Function -> Function

Returns the inverse of a function. This is a partial function.

evalTree :: (Num a, Floating a) => Vector a -> Vector Double -> (Double -> a) -> Fix SRTree -> a

Evaluates the tree given a vector of variable values, a vector of parameter values and a function that takes a Double and change to whatever type the variables have. This is useful when working with datasets of many values per variables.

relabelParams :: Fix SRTree -> Fix SRTree

Relabel the parameters incrementaly starting from 0

constsToParam :: Fix SRTree -> (Fix SRTree, [Double])

Change constant values to a parameter, returning the changed tree and a list of parameter values

floatConstsToParam :: Fix SRTree -> (Fix SRTree, [Double])

Same as constsToParam but does not change constant values that can be converted to integer without loss of precision