Data.SRTree.Datasets

Description

Utility library to handle regression datasets this module exports only the loadDataset function.

Documentation

loadDataset :: FilePath -> Bool -> IO ((SRMatrix, PVector, SRMatrix, PVector), String)

loadDataset loads a dataset with a filename in the format: filename.ext:start_row:end_row:target:features

where

*start_row:end_row** is the range of the training rows (default 0:nrows-1). every other row not included in this range will be used as validation

*target** is either the name of the PVector (if the datafile has headers) or the index of the target variable

*features** is a comma separated list of SRMatrix names or indices to be used as input variables of the regression model.