Data.SRTree.Eval

Description

Evaluation of SRTree expressions

Documentation

evalTree :: SRMatrix -> PVector -> Fix SRTree -> SRVector

Evaluates the tree given a vector of variable values, a vector of parameter values and a function that takes a Double and change to whatever type the variables have. This is useful when working with datasets of many values per variables.

evalOp :: Floating a => Op -> a -> a -> a

evalFun :: Floating a => Function -> a -> a

cbrt :: Floating a => a -> a

inverseFunc :: Function -> Function

Returns the inverse of a function. This is a partial function.

invertibles :: [Function]

List of invertible functions

evalInverse :: Floating a => Function -> a -> a

evals the inverse of a function

invright :: Floating a => Op -> a -> a -> a

evals the right inverse of an operator

invleft :: Floating a => Op -> a -> a -> a

evals the left inverse of an operator

replicateAs :: SRMatrix -> Double -> SRVector

type SRVector = Array D Ix1 Double

Vector of target values

type PVector = Array S Ix1 Double

Vector of parameter values. Needs to be strict to be readily accesible.

type SRMatrix = Array S Ix2 Double

Matrix of features values

Orphan instances

Index ix => Floating (Array D ix Double)

Instance details

Methods

pi :: Array D ix Double

exp :: Array D ix Double -> Array D ix Double

log :: Array D ix Double -> Array D ix Double

sqrt :: Array D ix Double -> Array D ix Double

(**) :: Array D ix Double -> Array D ix Double -> Array D ix Double

logBase :: Array D ix Double -> Array D ix Double -> Array D ix Double

sin :: Array D ix Double -> Array D ix Double

cos :: Array D ix Double -> Array D ix Double

tan :: Array D ix Double -> Array D ix Double

asin :: Array D ix Double -> Array D ix Double

acos :: Array D ix Double -> Array D ix Double

atan :: Array D ix Double -> Array D ix Double

sinh :: Array D ix Double -> Array D ix Double

cosh :: Array D ix Double -> Array D ix Double

tanh :: Array D ix Double -> Array D ix Double

asinh :: Array D ix Double -> Array D ix Double

acosh :: Array D ix Double -> Array D ix Double

atanh :: Array D ix Double -> Array D ix Double

log1p :: Array D ix Double -> Array D ix Double

expm1 :: Array D ix Double -> Array D ix Double

log1pexp :: Array D ix Double -> Array D ix Double

log1mexp :: Array D ix Double -> Array D ix Double

Index ix => Num (Array D ix Double)

Instance details

Methods

(+) :: Array D ix Double -> Array D ix Double -> Array D ix Double

(-) :: Array D ix Double -> Array D ix Double -> Array D ix Double

(*) :: Array D ix Double -> Array D ix Double -> Array D ix Double

negate :: Array D ix Double -> Array D ix Double

abs :: Array D ix Double -> Array D ix Double

signum :: Array D ix Double -> Array D ix Double

fromInteger :: Integer -> Array D ix Double

Index ix => Fractional (Array D ix Double)

Instance details

Methods

(/) :: Array D ix Double -> Array D ix Double -> Array D ix Double

recip :: Array D ix Double -> Array D ix Double

fromRational :: Rational -> Array D ix Double