Data.SRTree.Internal

Description

Expression tree for Symbolic Regression

Documentation

data SRTree ix val

Tree structure to be used with Symbolic Regression algorithms. This structure is parametrized by the indexing type to retrieve the values of a variable and the type of the output value.

Constructors

Empty

Var ix

Const val

Param ix

Fun Function (SRTree ix val)

Pow (SRTree ix val) Int

(SRTree ix val) `Add` (SRTree ix val)

(SRTree ix val) `Sub` (SRTree ix val)

(SRTree ix val) `Mul` (SRTree ix val)

(SRTree ix val) `Div` (SRTree ix val)

(SRTree ix val) `Power` (SRTree ix val)

(SRTree ix val) `LogBase` (SRTree ix val)

data Function

Functions that can be applied to a subtree.

Constructors

Id

Abs

Sin

Cos

Tan

Sinh

Cosh

Tanh

ASin

ACos

ATan

ASinh

ACosh

ATanh

Sqrt

Cbrt

Square

Log

Exp

class OptIntPow a where

A class for optimized (^^) operators for specific types. This was created because the integer power operator for interval arithmetic must be aware of the dependency problem, thus the default (^) doesn't work.

Methods

(^.) :: a -> Int -> a infix 8

traverseIx :: Applicative f => (ixa -> f ixb) -> SRTree ixa val -> f (SRTree ixb val)

Same as traverse but for the first type parameter.

arity :: SRTree ix val -> Int

Arity of the current node

getChildren :: SRTree ix val -> [SRTree ix val]

Get the children of a node. Returns an empty list in case of a leaf node.

countNodes :: SRTree ix val -> Int

Count the number of nodes in a tree.

countVarNodes :: SRTree ix val -> Int

Count the number of Var nodes

countOccurrences :: Eq ix => SRTree ix val -> ix -> Int

Count the occurrences of variable indexed as ix

deriveBy :: (Eq ix, Eq val, Floating val, OptIntPow val) => ix -> SRTree ix val -> SRTree ix val

Creates an SRTree representing the partial derivative of the input by the variable indexed by ix.

simplify :: (Eq ix, Eq val, Floating val, OptIntPow val) => SRTree ix val -> SRTree ix val

Simplifies the SRTree.

derivative :: (Eq ix, Eq val, Floating val) => Function -> SRTree ix val -> SRTree ix val

Derivative of a Function

evalFun :: Floating val => Function -> val -> val

Evaluates a function.

inverseFunc :: Function -> Function

Returns the inverse of a function. This is a partial function.

evalTree :: (Floating val, OptIntPow val) => SRTree ix val -> Reader (ix -> Maybe val) (Maybe val)

Evaluates a tree with the variables stored in a Reader monad.

evalTreeMap :: (Floating v1, OptIntPow v1, Floating v2, OptIntPow v2) => (v1 -> v2) -> SRTree ix v1 -> Reader (ix -> Maybe v2) (Maybe v2)

Evaluates a tree with the variables stored in a Reader monad while mapping the constant values to a different type.

evalTreeWithMap :: (Ord ix, Floating val, OptIntPow val) => SRTree ix val -> Map ix val -> Maybe val

Example of using evalTree with a Map.

evalTreeWithVector :: (Floating val, OptIntPow val) => SRTree Int val -> Vector val -> Maybe val

Example of using evalTree with a Vector.

relabelOccurrences :: forall ix val. Ord ix => SRTree ix val -> SRTree (ix, Int) val

Relabel occurences of a var into a tuple (ix, Int).

relabelParams :: Num ix => SRTree ix val -> SRTree ix val

Relabel the parameters sequentially starting from 0