Data.SRTree.Print

Description

Conversion functions to display the expression trees in different formats.

Documentation

data DisplayNodes ix val

Data structure containing the needed definitions to print a SRTree.

Constructors

D

Fields

_displayVar :: ix -> String

_displayPar :: ix -> String

_displayVal :: val -> String

_displayFun :: Function -> String

_displayPow :: String

_displayFloatPow :: String

showExpr :: (Show ix, Show val) => SRTree ix val -> DisplayNodes ix val -> String

Converts a tree to a String using the specifications given by DisplayNodes

showTree :: (Show ix, Show val) => SRTree ix val -> DisplayNodes ix val -> String

Converts a tree to a String using the specifications given by DisplayNodes

printExpr :: (Show ix, Show val) => SRTree ix val -> DisplayNodes ix val -> IO ()

Prints a tree as an expression using the specifications given by DisplayNodes

showDefault :: (Show ix, Show val) => SRTree ix val -> String

Displays a tree as an expression

showTikz :: (Show ix, Show val, RealFrac val) => SRTree ix val -> String

Displays a tree in Tikz format

showPython :: (Show ix, Show val) => SRTree ix val -> String

Displays a tree as a numpy compatible expression.

showLatex :: (Show ix, Show val) => SRTree ix val -> String

Displays a tree as a sympy compatible expression.