Data.SRTree.Random

Description

Functions to generate random trees and nodes.

Documentation

class HasVars p

Minimal complete definition

_vars

class HasVals p

Minimal complete definition

_range

class HasFuns p

Minimal complete definition

_funs

type HasEverything p = (HasVars p, HasVals p, HasExps p, HasFuns p)

Constraint synonym for all properties.

data FullParams ix val

A structure with every property

Constructors

P [ix] (val, val) (Int, Int) [Function]

type RndTree p ix val = ReaderT (p ix val) (StateT StdGen IO) (SRTree ix val)

RndTree is a Monad Transformer to generate random trees of type `SRTree ix val` given the parameters `p ix val` using the random number generator StdGen.

randomVar :: HasVars p => RndTree p ix val

Returns a random variable, the parameter p must have the HasVars property

randomConst :: (Ord val, Random val, HasVals p) => RndTree p ix val

Returns a random constant, the parameter p must have the HasConst property

randomPow :: (Ord val, Random val, HasExps p) => RndTree p ix val

Returns a random integer power node, the parameter p must have the HasExps property

randomFunction :: HasFuns p => RndTree p ix val

Returns a random function, the parameter p must have the HasFuns property

randomNode :: (Ord val, Random val, HasEverything p) => RndTree p ix val

Returns a random node, the parameter p must have every property.

randomNonTerminal :: (Ord val, Random val, HasEverything p) => RndTree p ix val

Returns a random non-terminal node, the parameter p must have every property.

randomTree :: (Ord val, Random val, HasEverything p) => Int -> RndTree p ix val

Returns a random tree with a limited budget, the parameter p must have every property.

randomTreeBalanced :: (Ord val, Random val, HasEverything p) => Int -> RndTree p ix val

Returns a random tree with a approximately a number n of nodes, the parameter p must have every property.